srcs/requirements.txt
Flask==2.0.3
Faker==33.0.0
psycopg==3.2.3               # Para interactuar con PostgreSQL sin ORM
python-dotenv==1.0.0         # Para gestionar las variables de entorno
Flask-SocketIO==5.3.0        # Para el chat en tiempo real
Flask-Mail==0.9.1            # Para enviar correos electrónicos
Flask-Login==0.6.2           # Para la gestión de sesiones de usuario
Flask-WTF==1.1.1             # Para formularios con validación
gunicorn==20.1.0             # Servidor WSGI para producción
requests==2.28.0             # Para realizar peticiones HTTP externas
werkzeug==2.0.3              # Librería que Flask usa internamente para utilidades
flask_jwt_extended==4.4.4  # O la versión que necesites
psycopg[binary]>=3.0,<4.0
psycopg_pool
psycopg2